task 2 completed

+++ Level 1 basics/2 oop/Task 2/B/Memo and Momo/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         System.out.println("Input");
 
         Scanner inputScan = new Scanner(System.in);
-        MemoAndMomo winner = new MemoAndMomo(inputScan.nextDouble() , inputScan.nextDouble(), inputScan.nextDouble());
+        MemoAndMomo winner = new MemoAndMomo(inputScan.nextLong() , inputScan.nextLong(), inputScan.nextLong());
 
         System.out.println("Output");
         System.out.println(winner.getWinner());
