+++ Level 1 basics/2 oop/Task 2/D/Ali Baba and Puzzles/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
         System.out.println("Input");
 
         Scanner inputScan = new Scanner(System.in);
-        AliBabaAndPuzzles checkpuzzle = new AliBabaAndPuzzles(inputScan.nextInt(),inputScan.nextInt(),inputScan.nextInt(),inputScan.nextInt());
+        AliBabaAndPuzzles checkpuzzle = new AliBabaAndPuzzles(inputScan.nextLong(),inputScan.nextLong(),inputScan.nextLong(),inputScan.nextLong());
 
         System.out.println("Output");
         checkpuzzle.solvePuzzle();
