+++ Level 1 basics/2 oop/Task 2/H/Data Type Guessing/src/Main.java
@@ -7,9 +7,9 @@ public class Main {
         Scanner inputScan = new Scanner(System.in);
 
         DataTypeGuessing checkType = new DataTypeGuessing();
-        checkType.n = inputScan.nextDouble();
-        checkType.k = inputScan.nextInt();
-        checkType.a = inputScan.nextInt();
+        checkType.n = inputScan.nextLong();
+        checkType.k = inputScan.nextLong();
+        checkType.a = inputScan.nextLong();
 
         System.out.println("Output");
         System.out.println(checkType.getDataType());
